+++ Packages/UGF.Code.Generate/Editor/Container.Info/ICodeGenerateContainerInfo.cs
@@ -1,25 +1,18 @@
 using System;
 
-namespace UGF.Code.Generate.Editor.Container.External
+namespace UGF.Code.Generate.Editor.Container.Info
 {
     /// <summary>
     /// Represents container external type information.
     /// </summary>
-    public interface ICodeGenerateContainerExternalInfo
+    public interface ICodeGenerateContainerInfo
     {
         /// <summary>
         /// Tries to get member info by the specified name.
         /// </summary>
         /// <param name="name">The name of the member to find.</param>
         /// <param name="member">The found member.</param>
-        bool TryGetMember<T>(string name, out T member) where T : CodeGenerateContainerExternalMemberInfo;
-
-        /// <summary>
-        /// Tries to get member info by the specified name.
-        /// </summary>
-        /// <param name="name">The name of the member to find.</param>
-        /// <param name="member">The found member.</param>
-        bool TryGetMember(string name, out CodeGenerateContainerExternalMemberInfo member);
+        bool TryGetMember(string name, out CodeGenerateContainerInfo.MemberInfo member);
 
         /// <summary>
         /// Tries to get target type.